WMC est fait, refactor du code Metrics, CSVWriter et JavaFile pour traiter une seulle classe interne.

+++ TP1/src/Class.java
@@ -75,7 +75,7 @@ public class Class {
         return BC;
     }
 
-    public void setBC(int BC) {
+    public void setBC(float BC) {
         this.BC = BC;
     }
 }
